+++ spec/dummy/app/controllers/application_controller.rb
@@ -1,7 +1,0 @@
-class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception
-
-  def current_user
-    'Test User'
-  end
-end