+++ spec/dummy/app/models/message.rb
@@ -1,3 +1,0 @@
-class Message < ApplicationRecord
-  rate_limit on: :ip_address, interval: 1.minute
-end