+++ test/dummy_app/app/controllers/concerns/require_current_user.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RequireCurrentUser
+  extend ActiveSupport::Concern
+
+  def require_current_user
+    return if current_user.present?
+
+    flash[:error] = 'You must be logged in to perform this action.'
+
+    redirect_to root_url
+  end
+end