+++ test/dummy_app/app/models/message.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class Message < ApplicationRecord
+  belongs_to :user
+
+  validates :subject, presence: true
+  validates :body, presence: true
+  validates :ip_address, presence: true
+
+  rate_limit on: :ip_address, interval: 1.minute
+end