+++ test/dummy_app/app/models/user.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class User < ApplicationRecord
+  has_many :messages
+
+  validates :email, presence: true
+end